+++ src/export_roms/test_multiple_tracks_chd_to_zso_should_do_nothing.rs
@@ -3,7 +3,7 @@ use super::super::import_dats;
 use super::super::import_roms;
 use super::*;
 use std::env;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use tempfile::{NamedTempFile, TempDir};
 use tokio::fs;
 
@@ -33,20 +33,27 @@ async fn test() {
     let tmp_directory = set_tmp_directory(PathBuf::from(tmp_directory.path()));
 
     let matches = import_dats::subcommand()
-        .get_matches_from(&["import-dats", "tests/Test System (20230618) (RVZ).dat"]);
+        .get_matches_from(&["import-dats", "tests/Test System (20200721).dat"]);
     import_dats::main(&mut connection, &matches, &progress_bar)
         .await
         .unwrap();
 
-    let romfile_path = tmp_directory.join("Test Game (USA).iso");
-    fs::copy(test_directory.join("Test Game (USA).iso"), &romfile_path)
-        .await
-        .unwrap();
+    let romfile_path = tmp_directory.join("Test Game (USA, Europe) (Multiple Tracks).cue");
+    fs::copy(
+        test_directory.join("Test Game (USA, Europe) (Multiple Tracks).cue"),
+        &romfile_path,
+    )
+    .await
+    .unwrap();
+    let romfile_path = tmp_directory.join("Test Game (USA, Europe) (Multiple Tracks).chd");
+    fs::copy(
+        test_directory.join("Test Game (USA, Europe) (Multiple Tracks).chd"),
+        &romfile_path,
+    )
+    .await
+    .unwrap();
 
     let system = find_systems(&mut connection).await.remove(0);
-    let system_directory = get_system_directory(&mut connection, &system)
-        .await
-        .unwrap();
 
     let matches = import_roms::subcommand()
         .get_matches_from(&["import-roms", romfile_path.as_os_str().to_str().unwrap()]);
@@ -54,47 +61,33 @@ async fn test() {
         .await
         .unwrap();
 
-    let roms = find_roms_with_romfile_by_system_id(&mut connection, system.id).await;
-    let romfile = find_romfile_by_id(&mut connection, roms[0].romfile_id.unwrap()).await;
     let mut roms_by_game_id: HashMap<i64, Vec<Rom>> = HashMap::new();
-    roms_by_game_id.insert(roms[0].game_id, roms);
     let mut romfiles_by_id: HashMap<i64, Romfile> = HashMap::new();
-    romfiles_by_id.insert(romfile.id, romfile);
+    let roms = find_roms_with_romfile_by_system_id(&mut connection, system.id).await;
+    for rom in &roms {
+        let romfile = find_romfile_by_id(&mut connection, rom.romfile_id.unwrap()).await;
+        romfiles_by_id.insert(romfile.id, romfile);
+    }
+    roms_by_game_id.insert(roms[0].game_id, roms);
+
+    let destination_directory = tmp_directory.join("destination");
+    create_directory(&progress_bar, &destination_directory, true)
+        .await
+        .unwrap();
 
     // when
-    to_rvz(
+    to_zso(
         &mut connection,
         &progress_bar,
+        &destination_directory,
         roms_by_game_id,
         romfiles_by_id,
-        true,
-        true,
-        &HashAlgorithm::Crc,
-        &RvzCompressionAlgorithm::Zstd,
-        5,
-        128,
     )
     .await
     .unwrap();
 
     // then
-    let roms = find_roms_with_romfile_by_system_id(&mut connection, system.id).await;
-    assert_eq!(roms.len(), 1);
-    let romfiles = find_romfiles(&mut connection).await;
-    assert_eq!(romfiles.len(), 1);
-
-    let rom = roms.first().unwrap();
-    assert_eq!(rom.name, "Test Game (USA).iso");
-
-    let romfile = romfiles.first().unwrap();
-    assert_eq!(
-        romfile.path,
-        system_directory
-            .join("Test Game (USA).rvz")
-            .as_os_str()
-            .to_str()
-            .unwrap(),
-    );
-    assert!(Path::new(&romfile.path).is_file());
-    assert_eq!(rom.romfile_id, Some(romfile.id));
+    assert!(!destination_directory
+        .join("Test Game (USA, Europe).zso")
+        .is_file());
 }