+++ src/export_roms/test_original_to_zip_with_incorrect_name.rs
@@ -2,8 +2,7 @@ use super::super::database::*;
 use super::super::import_dats;
 use super::super::import_roms;
 use super::*;
-use std::env;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use tempfile::{NamedTempFile, TempDir};
 use tokio::fs;
 
@@ -13,14 +12,6 @@ async fn test() {
     let _guard = MUTEX.lock().await;
 
     let test_directory = Path::new("tests");
-    env::set_var(
-        "PATH",
-        format!(
-            "{}:{}",
-            test_directory.as_os_str().to_str().unwrap(),
-            env::var("PATH").unwrap()
-        ),
-    );
     let progress_bar = ProgressBar::hidden();
 
     let db_file = NamedTempFile::new().unwrap();
@@ -33,20 +24,20 @@ async fn test() {
     let tmp_directory = set_tmp_directory(PathBuf::from(tmp_directory.path()));
 
     let matches = import_dats::subcommand()
-        .get_matches_from(&["import-dats", "tests/Test System (20230618) (RVZ).dat"]);
+        .get_matches_from(&["import-dats", "tests/Test System (20200721).dat"]);
     import_dats::main(&mut connection, &matches, &progress_bar)
         .await
         .unwrap();
 
-    let romfile_path = tmp_directory.join("Test Game (USA).iso");
-    fs::copy(test_directory.join("Test Game (USA).iso"), &romfile_path)
-        .await
-        .unwrap();
+    let romfile_path = tmp_directory.join("Test Game (USA, Europe).rom");
+    fs::copy(
+        test_directory.join("Test Game (USA, Europe).rom"),
+        &romfile_path,
+    )
+    .await
+    .unwrap();
 
     let system = find_systems(&mut connection).await.remove(0);
-    let system_directory = get_system_directory(&mut connection, &system)
-        .await
-        .unwrap();
 
     let matches = import_roms::subcommand()
         .get_matches_from(&["import-roms", romfile_path.as_os_str().to_str().unwrap()]);
@@ -61,40 +52,30 @@ async fn test() {
     let mut romfiles_by_id: HashMap<i64, Romfile> = HashMap::new();
     romfiles_by_id.insert(romfile.id, romfile);
 
+    let destination_directory = tmp_directory.join("destination");
+    create_directory(&progress_bar, &destination_directory, true)
+        .await
+        .unwrap();
+
+    let matches = subcommand().get_matches_from(&[
+        "convert-roms",
+        "-f",
+        "ZIP",
+        "-s",
+        "Test System",
+        "-n",
+        "test gqme",
+        "-d",
+        destination_directory.as_os_str().to_str().unwrap(),
+    ]);
+
     // when
-    to_rvz(
-        &mut connection,
-        &progress_bar,
-        roms_by_game_id,
-        romfiles_by_id,
-        true,
-        true,
-        &HashAlgorithm::Crc,
-        &RvzCompressionAlgorithm::Zstd,
-        5,
-        128,
-    )
-    .await
-    .unwrap();
+    main(&mut connection, &matches, &progress_bar)
+        .await
+        .unwrap();
 
     // then
-    let roms = find_roms_with_romfile_by_system_id(&mut connection, system.id).await;
-    assert_eq!(roms.len(), 1);
-    let romfiles = find_romfiles(&mut connection).await;
-    assert_eq!(romfiles.len(), 1);
-
-    let rom = roms.first().unwrap();
-    assert_eq!(rom.name, "Test Game (USA).iso");
-
-    let romfile = romfiles.first().unwrap();
-    assert_eq!(
-        romfile.path,
-        system_directory
-            .join("Test Game (USA).rvz")
-            .as_os_str()
-            .to_str()
-            .unwrap(),
-    );
-    assert!(Path::new(&romfile.path).is_file());
-    assert_eq!(rom.romfile_id, Some(romfile.id));
+    assert!(!destination_directory
+        .join("Test Game (USA, Europe).zip")
+        .is_file());
 }