+++ src/export_roms/test_rvz_to_iso.rs
@@ -3,7 +3,7 @@ use super::super::import_dats;
 use super::super::import_roms;
 use super::*;
 use std::env;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use tempfile::{NamedTempFile, TempDir};
 use tokio::fs;
 
@@ -38,15 +38,12 @@ async fn test() {
         .await
         .unwrap();
 
-    let romfile_path = tmp_directory.join("Test Game (USA).iso");
-    fs::copy(test_directory.join("Test Game (USA).iso"), &romfile_path)
+    let romfile_path = tmp_directory.join("Test Game (USA).rvz");
+    fs::copy(test_directory.join("Test Game (USA).rvz"), &romfile_path)
         .await
         .unwrap();
 
     let system = find_systems(&mut connection).await.remove(0);
-    let system_directory = get_system_directory(&mut connection, &system)
-        .await
-        .unwrap();
 
     let matches = import_roms::subcommand()
         .get_matches_from(&["import-roms", romfile_path.as_os_str().to_str().unwrap()]);
@@ -54,6 +51,8 @@ async fn test() {
         .await
         .unwrap();
 
+    let games = find_games_with_romfiles_by_system_id(&mut connection, system.id).await;
+    let games_by_id: HashMap<i64, Game> = games.into_iter().map(|game| (game.id, game)).collect();
     let roms = find_roms_with_romfile_by_system_id(&mut connection, system.id).await;
     let romfile = find_romfile_by_id(&mut connection, roms[0].romfile_id.unwrap()).await;
     let mut roms_by_game_id: HashMap<i64, Vec<Rom>> = HashMap::new();
@@ -61,40 +60,23 @@ async fn test() {
     let mut romfiles_by_id: HashMap<i64, Romfile> = HashMap::new();
     romfiles_by_id.insert(romfile.id, romfile);
 
+    let destination_directory = tmp_directory.join("destination");
+    create_directory(&progress_bar, &destination_directory, true)
+        .await
+        .unwrap();
+
     // when
-    to_rvz(
-        &mut connection,
+    to_original(
         &progress_bar,
+        &destination_directory,
+        &system,
+        games_by_id,
         roms_by_game_id,
         romfiles_by_id,
-        true,
-        true,
-        &HashAlgorithm::Crc,
-        &RvzCompressionAlgorithm::Zstd,
-        5,
-        128,
     )
     .await
     .unwrap();
 
     // then
-    let roms = find_roms_with_romfile_by_system_id(&mut connection, system.id).await;
-    assert_eq!(roms.len(), 1);
-    let romfiles = find_romfiles(&mut connection).await;
-    assert_eq!(romfiles.len(), 1);
-
-    let rom = roms.first().unwrap();
-    assert_eq!(rom.name, "Test Game (USA).iso");
-
-    let romfile = romfiles.first().unwrap();
-    assert_eq!(
-        romfile.path,
-        system_directory
-            .join("Test Game (USA).rvz")
-            .as_os_str()
-            .to_str()
-            .unwrap(),
-    );
-    assert!(Path::new(&romfile.path).is_file());
-    assert_eq!(rom.romfile_id, Some(romfile.id));
+    assert!(destination_directory.join("Test Game (USA).iso").is_file());
 }