+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Add a `-s` flag to `convert-roms` to select systems by name 
 - Update the `-s` flag in `import-roms` to behave the same as in `convert-roms`
 - Add a `-f` flag in `purge-roms` to track and delete foreign files in the ROM directory
+- Add a new `export-roms` subcommand
+- Update 
 
 # 0.18.0